# payment-api/generated/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    This is a sample API for managing payments and balance.
    It includes endpoints for making payments and checking the account balance.
  title: Sample API
  version: 1.0.0
servers:
- description: Local server for testing goals only
  url: http://localhost:8080
tags:
- description: Operations related to payments
  name: Payment
- description: Operations related to balance
  name: Balance
paths:
  /pay/{amount}:
    get:
      description: "This is request to subtract the amount from the account. \nIf\
        \ the amount is greater than the balance, it will return false.\n"
      operationId: makePayment
      parameters:
      - description: Amount to be paid
        example:
          amount: 10.3
        explode: false
        in: path
        name: amount
        required: true
        schema:
          format: float
          minimum: 0
          type: number
        style: simple
      responses:
        "200":
          content:
            application/json:
              examples:
                Success example:
                  value:
                    success: true
                    currentBalance: 90.2
                    error: null
              schema:
                $ref: "#/components/schemas/UpdateBalanceResponse"
          description: |
            Payment information
            This will return true if the payment was successful, false otherwise.
        "400":
          content:
            application/json:
              examples:
                Low balance error example:
                  value:
                    success: false
                    currentBalance: 10.3
                    error: Insufficient balance
                Validation error example:
                  value:
                    success: false
                    currentBalance: 10.3
                    error: Validation exception
              schema:
                $ref: "#/components/schemas/UpdateBalanceResponse"
          description: Bad request in case of invalid amount
      summary: Making a payment
      tags:
      - Payment
      x-accepts:
      - application/json
      x-tags:
      - tag: Payment
  /balance:
    get:
      operationId: getBalance
      responses:
        "200":
          content:
            application/json:
              example:
                balance: 100.5
              schema:
                $ref: "#/components/schemas/GetBalanceResponse"
          description: Balance information
      summary: Get the balance on the account
      tags:
      - Balance
      x-accepts:
      - application/json
      x-tags:
      - tag: Balance
components:
  schemas:
    UpdateBalanceResponse:
      example:
        success: true
        currentBalance: 0.8008282
        error: error
      properties:
        success:
          type: boolean
        currentBalance:
          format: float
          type: number
        error:
          type: string
      type: object
    GetBalanceResponse:
      example:
        balance: 0.8008282
      properties:
        balance:
          format: float
          type: number
      type: object